+++ src/yyjson.c
@@ -197,18 +197,6 @@ yyjson_api uint32_t yyjson_version(void) {
 #   define YYJSON_DOUBLE_MATH_CORRECT 1
 #endif
 
-/*
- Microsoft Visual C++ 6.0 doesn't support converting number from u64 to f64:
- error C2520: conversion from unsigned __int64 to double not implemented.
- */
-#ifndef YYJSON_U64_TO_F64_NO_IMPL
-#   if (0 < YYJSON_MSC_VER) && (YYJSON_MSC_VER <= 1200)
-#       define YYJSON_U64_TO_F64_NO_IMPL 1
-#   else
-#       define YYJSON_U64_TO_F64_NO_IMPL 0
-#   endif
-#endif
-
 /* endian */
 #if yyjson_has_include(<sys/types.h>)
 #    include <sys/types.h>
@@ -412,6 +400,9 @@ yyjson_api uint32_t yyjson_version(void) {
 #undef  U64
 #define U64(hi, lo) ((((u64)hi##UL) << 32U) + lo##UL)
 
+/* Used to cast away (remove) const qualifier. */
+#define constcast(type) (type)(void *)(size_t)(const void *)
+
 
 
 /*==============================================================================
@@ -530,7 +521,7 @@ typedef union v64_uni { v64 v; u64 u; } v64_uni;
 
 static_inline void byte_move_2(void *dst, const void *src) {
 #if YYJSON_DISABLE_UNALIGNED_MEMORY_ACCESS
-    repeat2_incr(byte_move_idx);
+    repeat2_incr(byte_move_idx)
 #else
     memmove(dst, src, 2);
 #endif
@@ -538,7 +529,7 @@ static_inline void byte_move_2(void *dst, const void *src) {
 
 static_inline void byte_move_4(void *dst, const void *src) {
 #if YYJSON_DISABLE_UNALIGNED_MEMORY_ACCESS
-    repeat4_incr(byte_move_idx);
+    repeat4_incr(byte_move_idx)
 #else
     memmove(dst, src, 4);
 #endif
@@ -546,7 +537,7 @@ static_inline void byte_move_4(void *dst, const void *src) {
 
 static_inline void byte_move_8(void *dst, const void *src) {
 #if YYJSON_DISABLE_UNALIGNED_MEMORY_ACCESS
-    repeat8_incr(byte_move_idx);
+    repeat8_incr(byte_move_idx)
 #else
     memmove(dst, src, 8);
 #endif
@@ -554,7 +545,7 @@ static_inline void byte_move_8(void *dst, const void *src) {
 
 static_inline void byte_move_16(void *dst, const void *src) {
 #if YYJSON_DISABLE_UNALIGNED_MEMORY_ACCESS
-    repeat16_incr(byte_move_idx);
+    repeat16_incr(byte_move_idx)
 #else
     memmove(dst, src, 16);
 #endif
@@ -562,7 +553,7 @@ static_inline void byte_move_16(void *dst, const void *src) {
 
 static_inline void byte_copy_2(void *dst, const void *src) {
 #if YYJSON_DISABLE_UNALIGNED_MEMORY_ACCESS
-    repeat2_incr(byte_move_idx);
+    repeat2_incr(byte_move_idx)
 #else
     memcpy(dst, src, 2);
 #endif
@@ -570,7 +561,7 @@ static_inline void byte_copy_2(void *dst, const void *src) {
 
 static_inline void byte_copy_4(void *dst, const void *src) {
 #if YYJSON_DISABLE_UNALIGNED_MEMORY_ACCESS
-    repeat4_incr(byte_move_idx);
+    repeat4_incr(byte_move_idx)
 #else
     memcpy(dst, src, 4);
 #endif
@@ -578,7 +569,7 @@ static_inline void byte_copy_4(void *dst, const void *src) {
 
 static_inline void byte_copy_8(void *dst, const void *src) {
 #if YYJSON_DISABLE_UNALIGNED_MEMORY_ACCESS
-    repeat8_incr(byte_move_idx);
+    repeat8_incr(byte_move_idx)
 #else
     memcpy(dst, src, 8);
 #endif
@@ -586,7 +577,7 @@ static_inline void byte_copy_8(void *dst, const void *src) {
 
 static_inline void byte_copy_16(void *dst, const void *src) {
 #if YYJSON_DISABLE_UNALIGNED_MEMORY_ACCESS
-    repeat16_incr(byte_move_idx);
+    repeat16_incr(byte_move_idx)
 #else
     memcpy(dst, src, 16);
 #endif
@@ -622,21 +613,21 @@ static_inline bool byte_match_4(void *buf, const char *pat) {
 
 static_inline u16 byte_load_2(const void *src) {
     v16_uni uni;
-    uni.v = *(v16 *)src;
+    uni.v = *(const v16 *)src;
     return uni.u;
 }
 
 static_inline u32 byte_load_3(const void *src) {
     v32_uni uni;
-    ((v16_uni *)&uni)->v = *(v16 *)src;
-    uni.v.c3 = ((char *)src)[2];
+    ((v16_uni *)&uni)->v = *(const v16 *)src;
+    uni.v.c3 = ((const char *)src)[2];
     uni.v.c4 = 0;
     return uni.u;
 }
 
 static_inline u32 byte_load_4(const void *src) {
     v32_uni uni;
-    uni.v = *(v32 *)src;
+    uni.v = *(const v32 *)src;
     return uni.u;
 }
 
@@ -729,8 +720,7 @@ static_inline f64 normalized_u64_to_f64(u64 val) {
 
 /** Returns whether the size is overflow after increment. */
 static_inline bool size_add_is_overflow(usize size, usize add) {
-    usize val = size + add;
-    return (val < size) | (val < add);
+    return size > (size + add);
 }
 
 /** Returns whether the size is power of 2 (size should not be 0). */
@@ -1437,7 +1427,7 @@ static usize yyjson_imut_copy(yyjson_val **val_ptr, char **buf_ptr,
     } else if (type == YYJSON_TYPE_STR || type == YYJSON_TYPE_RAW) {
         char *buf = *buf_ptr;
         usize len = unsafe_yyjson_get_len(mval);
-        memcpy((void *)buf, (void *)mval->uni.str, len);
+        memcpy((void *)buf, (const void *)mval->uni.str, len);
         buf[len] = '\0';
         val->tag = mval->tag;
         val->uni.str = buf;
@@ -1479,7 +1469,7 @@ yyjson_api yyjson_doc *yyjson_mut_val_imut_copy(yyjson_mut_val *mval,
     doc = (yyjson_doc *)alc->malloc(alc->ctx, buf_size);
     if (!doc) return NULL;
     memset(doc, 0, sizeof(yyjson_doc));
-    val_hdr = (yyjson_val *)((char *)(void *)doc + hdr_size);
+    val_hdr = (yyjson_val *)(void *)((char *)(void *)doc + hdr_size);
     doc->root = val_hdr;
     doc->alc = *alc;
     
@@ -3135,7 +3125,7 @@ static_noinline bool skip_spaces_and_comments(u8 **ptr) {
 
 /**
  Check truncated string.
- Returns true if `cur` match `lit` but is truncated.
+ Returns true if `cur` match `str` but is truncated.
  */
 static_inline bool is_truncated_str(u8 *cur, u8 *end,
                                     const char *str,
@@ -3208,34 +3198,33 @@ static_noinline bool is_truncated_end(u8 *hdr, u8 *cur, u8 *end,
             u8 c0 = cur[0], c1 = cur[1], c2 = cur[2];
             if (len == 1) {
                 /* 2 bytes UTF-8, truncated */
-                if ((c0 & 0xE0) == 0xC0 && (c0 & 0x1E)) return true;
+                if ((c0 & 0xE0) == 0xC0 && (c0 & 0x1E) != 0x00) return true;
                 /* 3 bytes UTF-8, truncated */
                 if ((c0 & 0xF0) == 0xE0) return true;
                 /* 4 bytes UTF-8, truncated */
-                if ((c0 & 0xF8) == 0xF0) return true;
+                if ((c0 & 0xF8) == 0xF0 && (c0 & 0x07) <= 0x04) return true;
             }
             if (len == 2) {
                 /* 3 bytes UTF-8, truncated */
                 if ((c0 & 0xF0) == 0xE0 &&
-                    (c1 & 0xC0) == 0x80 &&
-                    !((c0 & 0x0F) == 0x00 && (c1 & 0x20) == 0x00) &&
-                    !((c0 & 0x0F) == 0x0D && (c1 & 0x20) == 0x20)) {
-                    return true;
+                    (c1 & 0xC0) == 0x80) {
+                    u8 pat = (u8)(((c0 & 0x0F) << 1) | ((c1 & 0x20) >> 5));
+                    return 0x01 <= pat && pat != 0x1B;
                 }
                 /* 4 bytes UTF-8, truncated */
                 if ((c0 & 0xF8) == 0xF0 &&
-                    (c1 & 0xC0) == 0x80 &&
-                    !((c0 & 0x07) == 0x00 && (c1 & 0x30) == 0x00)) {
-                    return true;
+                    (c1 & 0xC0) == 0x80) {
+                    u8 pat = (u8)(((c0 & 0x07) << 2) | ((c1 & 0x30) >> 4));
+                    return 0x01 <= pat && pat <= 0x10;
                 }
             }
             if (len == 3) {
                 /* 4 bytes UTF-8, truncated */
                 if ((c0 & 0xF8) == 0xF0 &&
                     (c1 & 0xC0) == 0x80 &&
-                    (c2 & 0xC0) == 0x80 &&
-                    !((c0 & 0x07) == 0x00 && (c1 & 0x30) == 0x00)) {
-                    return true;
+                    (c2 & 0xC0) == 0x80) {
+                    u8 pat = (u8)(((c0 & 0x07) << 2) | ((c1 & 0x30) >> 4));
+                    return 0x01 <= pat && pat <= 0x10;
                 }
             }
         }
@@ -3541,6 +3530,12 @@ static_inline bool read_number(u8 **ptr,
     return false; \
 } while (false)
     
+#define return_0() do { \
+    val->tag = YYJSON_TYPE_NUM | (u8)((u8)sign << 3); \
+    val->uni.u64 = 0; \
+    *end = cur; return true; \
+} while (false)
+
 #define return_i64(_v) do { \
     val->tag = YYJSON_TYPE_NUM | (u8)((u8)sign << 3); \
     val->uni.u64 = (u64)(sign ? (u64)(~(_v) + 1) : (u64)(_v)); \
@@ -3602,7 +3597,7 @@ static_inline bool read_number(u8 **ptr,
             return_err(cur, "no digit after minus sign");
         }
         /* begin with 0 */
-        if (likely(!digi_is_digit_or_fp(*++cur))) return_i64(0);
+        if (likely(!digi_is_digit_or_fp(*++cur))) return_0();
         if (likely(*cur == '.')) {
             dot_pos = cur++;
             if (unlikely(!digi_is_digit(*cur))) {
@@ -3644,7 +3639,7 @@ static_inline bool read_number(u8 **ptr,
 #define expr_intg(i) \
     if (likely((num = (u64)(cur[i] - (u8)'0')) <= 9)) sig = num + sig * 10; \
     else { goto digi_sepr_##i; }
-    repeat_in_1_18(expr_intg);
+    repeat_in_1_18(expr_intg)
 #undef expr_intg
     
     
@@ -4105,6 +4100,7 @@ digi_finish:
 #undef has_flag
 #undef return_err
 #undef return_inf
+#undef return_0
 #undef return_i64
 #undef return_f64
 #undef return_f64_raw
@@ -4131,6 +4127,12 @@ static_noinline bool read_number(u8 **ptr,
     return false; \
 } while (false)
     
+#define return_0() do { \
+    val->tag = YYJSON_TYPE_NUM | (u64)((u8)sign << 3); \
+    val->uni.u64 = 0; \
+    *end = cur; return true; \
+} while (false)
+
 #define return_i64(_v) do { \
     val->tag = YYJSON_TYPE_NUM | (u64)((u8)sign << 3); \
     val->uni.u64 = (u64)(sign ? (u64)(~(_v) + 1) : (u64)(_v)); \
@@ -4181,7 +4183,7 @@ static_noinline bool read_number(u8 **ptr,
         if (unlikely(digi_is_digit(*cur))) {
             return_err(cur - 1, "number with leading zero is not allowed");
         }
-        if (!digi_is_fp(*cur)) return_i64(0);
+        if (!digi_is_fp(*cur)) return_0();
         goto read_double;
     }
     
@@ -4189,7 +4191,7 @@ static_noinline bool read_number(u8 **ptr,
 #define expr_intg(i) \
     if (likely((num = (u64)(cur[i] - (u8)'0')) <= 9)) sig = num + sig * 10; \
     else { cur += i; goto intg_end; }
-    repeat_in_1_18(expr_intg);
+    repeat_in_1_18(expr_intg)
 #undef expr_intg
     
     /* here are 19 continuous digits, skip them */
@@ -4262,7 +4264,7 @@ read_double:
             return_err(hdr, "strtod() failed to parse the number");
         }
     }
-    if (unlikely(val->uni.f64 == HUGE_VAL || val->uni.f64 == -HUGE_VAL)) {
+    if (unlikely(val->uni.f64 >= HUGE_VAL || val->uni.f64 <= -HUGE_VAL)) {
         if (!ext) {
             return_err(hdr, "number is infinity when parsed as double");
         }
@@ -4273,6 +4275,7 @@ read_double:
     
 #undef has_flag
 #undef return_err
+#undef return_0
 #undef return_i64
 #undef return_f64
 }
@@ -4449,7 +4452,7 @@ skip_ascii_begin:
          while (true) repeat16({
             if (likely(!(char_is_ascii_stop(*src)))) src++;
             else break;
-         });
+         })
      */
 #define expr_jump(i) \
     if (likely(!char_is_ascii_stop(src[i]))) {} \
@@ -4460,10 +4463,10 @@ skip_ascii_begin:
     src += i; \
     goto skip_ascii_end;
     
-    repeat16_incr(expr_jump);
+    repeat16_incr(expr_jump)
     src += 16;
     goto skip_ascii_begin;
-    repeat16_incr(expr_stop);
+    repeat16_incr(expr_stop)
     
 #undef expr_jump
 #undef expr_stop
@@ -4594,7 +4597,7 @@ copy_ascii:
          while (true) repeat16({
             if (unlikely(char_is_ascii_stop(*src))) break;
             *dst++ = *src++;
-         });
+         })
      */
 #if YYJSON_IS_REAL_GCC
 #   define expr_jump(i) \
@@ -4605,7 +4608,7 @@ copy_ascii:
     if (likely(!(char_is_ascii_stop(src[i])))) {} \
     else { goto copy_ascii_stop_##i; }
 #endif
-    repeat16_incr(expr_jump);
+    repeat16_incr(expr_jump)
 #undef expr_jump
     
     byte_move_16(dst, src);
@@ -5370,12 +5373,12 @@ arr_val_begin:
     while (true) repeat16({
         if (byte_match_2(cur, "  ")) cur += 2;
         else break;
-    });
+    })
 #else
     while (true) repeat16({
         if (likely(byte_match_2(cur, "  "))) cur += 2;
         else break;
-    });
+    })
 #endif
     
     if (*cur == '{') {
@@ -5501,12 +5504,12 @@ obj_key_begin:
     while (true) repeat16({
         if (byte_match_2(cur, "  ")) cur += 2;
         else break;
-    });
+    })
 #else
     while (true) repeat16({
         if (likely(byte_match_2(cur, "  "))) cur += 2;
         else break;
-    });
+    })
 #endif
     if (likely(*cur == '"')) {
         val_incr();
@@ -5918,7 +5921,7 @@ const char *yyjson_read_number(const char *dat,
     return NULL; \
 } while (false)
     
-    u8 *hdr = (u8 *)dat, *cur = hdr;
+    u8 *hdr = constcast(u8 *)dat, *cur = hdr;
     bool raw; /* read number as raw */
     bool ext; /* allow inf and nan */
     u8 *raw_end; /* raw end for null-terminator */
@@ -6576,7 +6579,7 @@ static_noinline u8 *write_f64_raw(u8 *buf, u64 raw, yyjson_write_flag flg) {
      For non-IEEE formats, 17 is used to avoid buffer overflow,
      round-trip is not guaranteed.
      */
-#if defined(DBL_DECIMAL_DIG)
+#if defined(DBL_DECIMAL_DIG) && DBL_DECIMAL_DIG != 17
     int dig = DBL_DECIMAL_DIG > 17 ? 17 : DBL_DECIMAL_DIG;
 #else
     int dig = 17;
@@ -7026,26 +7029,26 @@ copy_ascii:
     cur += i; src += i; goto copy_utf8;
     
     while (end - src >= 16) {
-        repeat16_incr(expr_jump);
+        repeat16_incr(expr_jump)
         byte_copy_16(cur, src);
         cur += 16; src += 16;
     }
     
     while (end - src >= 4) {
-        repeat4_incr(expr_jump);
+        repeat4_incr(expr_jump)
         byte_copy_4(cur, src);
         cur += 4; src += 4;
     }
     
     while (end > src) {
-        expr_jump(0);
+        expr_jump(0)
         *cur++ = *src++;
     }
     
     *cur++ = '"';
     return cur;
     
-    repeat16_incr(expr_stop);
+    repeat16_incr(expr_stop)
     
 #undef expr_jump
 #undef expr_stop
@@ -7440,7 +7443,7 @@ static_inline u8 *yyjson_write_minify(const yyjson_val *root,
     ctx = (yyjson_write_ctx *)(void *)end;
     
 doc_begin:
-    val = (yyjson_val *)root;
+    val = constcast(yyjson_val *)root;
     val_type = unsafe_yyjson_get_type(val);
     ctn_obj = (val_type == YYJSON_TYPE_OBJ);
     ctn_len = unsafe_yyjson_get_len(val) << (u8)ctn_obj;
@@ -7613,7 +7616,7 @@ static_inline u8 *yyjson_write_pretty(const yyjson_val *root,
     ctx = (yyjson_write_ctx *)(void *)end;
     
 doc_begin:
-    val = (yyjson_val *)root;
+    val = constcast(yyjson_val *)root;
     val_type = unsafe_yyjson_get_type(val);
     ctn_obj = (val_type == YYJSON_TYPE_OBJ);
     ctn_len = unsafe_yyjson_get_len(val) << (u8)ctn_obj;
@@ -7755,7 +7758,7 @@ char *yyjson_val_write_opts(const yyjson_val *val,
     yyjson_write_err dummy_err;
     usize dummy_dat_len;
     yyjson_alc alc = alc_ptr ? *alc_ptr : YYJSON_DEFAULT_ALC;
-    yyjson_val *root = (yyjson_val *)val;
+    yyjson_val *root = constcast(yyjson_val *)val;
     
     err = err ? err : &dummy_err;
     dat_len = dat_len ? dat_len : &dummy_dat_len;
@@ -7798,7 +7801,7 @@ bool yyjson_val_write_file(const char *path,
     yyjson_write_err dummy_err;
     u8 *dat;
     usize dat_len = 0;
-    yyjson_val *root = (yyjson_val *)val;
+    yyjson_val *root = constcast(yyjson_val *)val;
     bool suc;
     
     alc_ptr = alc_ptr ? alc_ptr : &YYJSON_DEFAULT_ALC;
@@ -7922,7 +7925,7 @@ static_inline u8 *yyjson_mut_write_minify(const yyjson_mut_val *root,
     ctx = (yyjson_mut_write_ctx *)(void *)end;
     
 doc_begin:
-    val = (yyjson_mut_val *)root;
+    val = constcast(yyjson_mut_val *)root;
     val_type = unsafe_yyjson_get_type(val);
     ctn_obj = (val_type == YYJSON_TYPE_OBJ);
     ctn_len = unsafe_yyjson_get_len(val) << (u8)ctn_obj;
@@ -8100,7 +8103,7 @@ static_inline u8 *yyjson_mut_write_pretty(const yyjson_mut_val *root,
     ctx = (yyjson_mut_write_ctx *)(void *)end;
     
 doc_begin:
-    val = (yyjson_mut_val *)root;
+    val = constcast(yyjson_mut_val *)root;
     val_type = unsafe_yyjson_get_type(val);
     ctn_obj = (val_type == YYJSON_TYPE_OBJ);
     ctn_len = unsafe_yyjson_get_len(val) << (u8)ctn_obj;
@@ -8248,7 +8251,7 @@ char *yyjson_mut_val_write_opts(const yyjson_mut_val *val,
     yyjson_write_err dummy_err;
     usize dummy_dat_len;
     yyjson_alc alc = alc_ptr ? *alc_ptr : YYJSON_DEFAULT_ALC;
-    yyjson_mut_val *root = (yyjson_mut_val *)val;
+    yyjson_mut_val *root = constcast(yyjson_mut_val *)val;
     
     err = err ? err : &dummy_err;
     dat_len = dat_len ? dat_len : &dummy_dat_len;
@@ -8291,7 +8294,7 @@ bool yyjson_mut_val_write_file(const char *path,
     yyjson_write_err dummy_err;
     u8 *dat;
     usize dat_len = 0;
-    yyjson_mut_val *root = (yyjson_mut_val *)val;
+    yyjson_mut_val *root = constcast(yyjson_mut_val *)val;
     bool suc;
     
     alc_ptr = alc_ptr ? alc_ptr : &YYJSON_DEFAULT_ALC;